move handleSubmit to SwapButton
