+++ site/docs/pages/swap/swap.mdx
@@ -8,10 +8,10 @@ The `Swap` component is a comprehensive interface for users to execute token swa
 ## Usage
 
 ```tsx [code]
-<Swap account={account} onError={onError} onSuccess={onSuccess}>
+<Swap account={account}>
   <SwapAmountInputV2 label="Sell" token={fromToken} type="from" />
   <SwapAmountInputV2 label="Buy" token={toToken} type="to" />
-  <SwapButton />
+  <SwapButton onError={onError} onSuccess={onSuccess} />
 </Swap>
 ```
